+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/models/cache/gettables/returns/ConvertedPackagingCreditGettable.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables
+package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables.returns
 
 import play.api.libs.json.JsPath
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.Gettable
 
 case object ConvertedPackagingCreditGettable extends Gettable[BigDecimal] {
 
-  override def path: JsPath = JsPath \ "data" \ toString
+  override def path: JsPath = JsPath \ toString
 
   override def toString: String = "convertedPackagingCredit"
 }
