+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/models/cache/gettables/returns/ExportedPlasticPackagingWeightGettable.scala
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables
+package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables.returns
 
 import play.api.libs.json.JsPath
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.Gettable
 
-case object ImportedPlasticPackagingWeightGettable extends Gettable[Long] {
+case object ExportedPlasticPackagingWeightGettable extends Gettable[Long] {
 
-  override def path: JsPath = JsPath \ "data" \ toString
+  override def path: JsPath = JsPath \ toString
 
-  override def toString: String = "importedPlasticPackagingWeight"
-
+  override def toString: String = "exportedPlasticPackagingWeight"
 }
