+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/models/cache/gettables/returns/ManufacturedPlasticPackagingWeightGettable.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables
+package uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.gettables.returns
 
 import play.api.libs.json.JsPath
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.cache.Gettable
 
 case object ManufacturedPlasticPackagingWeightGettable extends Gettable[Long] {
 
-  override def path: JsPath = JsPath \ "data" \ toString
+  override def path: JsPath = JsPath \ toString
 
   override def toString: String = "manufacturedPlasticPackagingWeight"
 }
